Refactor song list and voting functionality
Introduced `SongListItem` and `SongVote` components to modularize and simplify song list and vote handling. Added support for mode-specific behavior (e.g., 'creator' vs 'normal') and cleaned up redundant logic across components. Enhanced the code structure for better scalability and maintainability.
